+++ tiled/server/authentication.py
@@ -552,13 +552,13 @@ async def check_scopes(
             raise HTTPException(
                 status_code=HTTP_401_UNAUTHORIZED,
                 detail=(
-                    "Not enough permissions. "
-                    f"Requires scopes {settings.authenticator.scopes}. "
+                    "Authenticator has scopes that the request does not have. "
+                    f"Authenticator has scopes {settings.authenticator.scopes}. "
                     f"Request had scopes {list(scopes)}"
                 ),
                 headers=headers_for_401(request, security_scopes),
             )
-    elif not set(security_scopes.scopes).issubset(scopes):
+    if not set(security_scopes.scopes).issubset(scopes):
         raise HTTPException(
             status_code=HTTP_401_UNAUTHORIZED,
             detail=(